third_party/blink/web_tests/http/tests/worklet/webexposed/global-interface-listing-animation-worklet-expected.txt

CONSOLE MESSAGE: This test logs exposed APIs from AnimationWorkletGlobalScope
CONSOLE MESSAGE: [INTERFACES]
CONSOLE MESSAGE: interface AnimationWorkletGlobalScope : WorkletGlobalScope
CONSOLE MESSAGE:     attribute @@toStringTag
CONSOLE MESSAGE:     method constructor
CONSOLE MESSAGE: interface ByteLengthQueuingStrategy
CONSOLE MESSAGE:     attribute @@toStringTag
CONSOLE MESSAGE:     getter highWaterMark
CONSOLE MESSAGE:     getter size
CONSOLE MESSAGE:     method constructor
CONSOLE MESSAGE: interface CountQueuingStrategy
CONSOLE MESSAGE:     attribute @@toStringTag
CONSOLE MESSAGE:     getter highWaterMark
CONSOLE MESSAGE:     getter size
CONSOLE MESSAGE:     method constructor
CONSOLE MESSAGE: interface ReadableByteStreamController
CONSOLE MESSAGE:     attribute @@toStringTag
CONSOLE MESSAGE:     getter byobRequest
CONSOLE MESSAGE:     getter desiredSize
CONSOLE MESSAGE:     method close
CONSOLE MESSAGE:     method constructor
CONSOLE MESSAGE:     method enqueue
CONSOLE MESSAGE:     method error
CONSOLE MESSAGE: interface ReadableStream
CONSOLE MESSAGE:     attribute @@toStringTag
CONSOLE MESSAGE:     getter locked
CONSOLE MESSAGE:     method cancel
CONSOLE MESSAGE:     method constructor
CONSOLE MESSAGE:     method getReader
CONSOLE MESSAGE:     method pipeThrough
CONSOLE MESSAGE:     method pipeTo
CONSOLE MESSAGE:     method tee
CONSOLE MESSAGE: interface ReadableStreamBYOBReader
CONSOLE MESSAGE:     attribute @@toStringTag
CONSOLE MESSAGE:     getter closed
CONSOLE MESSAGE:     method cancel
CONSOLE MESSAGE:     method constructor
CONSOLE MESSAGE:     method read
CONSOLE MESSAGE:     method releaseLock
CONSOLE MESSAGE: interface ReadableStreamBYOBRequest
CONSOLE MESSAGE:     attribute @@toStringTag
CONSOLE MESSAGE:     getter view
CONSOLE MESSAGE:     method constructor
CONSOLE MESSAGE:     method respond
CONSOLE MESSAGE:     method respondWithNewView
CONSOLE MESSAGE: interface ReadableStreamDefaultController
CONSOLE MESSAGE:     attribute @@toStringTag
CONSOLE MESSAGE:     getter desiredSize
CONSOLE MESSAGE:     method close
CONSOLE MESSAGE:     method constructor
CONSOLE MESSAGE:     method enqueue
CONSOLE MESSAGE:     method error
CONSOLE MESSAGE: interface ReadableStreamDefaultReader
CONSOLE MESSAGE:     attribute @@toStringTag
CONSOLE MESSAGE:     getter closed
CONSOLE MESSAGE:     method cancel
CONSOLE MESSAGE:     method constructor
CONSOLE MESSAGE:     method read
CONSOLE MESSAGE:     method releaseLock
CONSOLE MESSAGE: interface TransformStream
CONSOLE MESSAGE:     attribute @@toStringTag
CONSOLE MESSAGE:     getter readable
CONSOLE MESSAGE:     getter writable
CONSOLE MESSAGE:     method constructor
CONSOLE MESSAGE: interface TransformStreamDefaultController
CONSOLE MESSAGE:     attribute @@toStringTag
CONSOLE MESSAGE:     getter desiredSize
CONSOLE MESSAGE:     method constructor
CONSOLE MESSAGE:     method enqueue
CONSOLE MESSAGE:     method error
CONSOLE MESSAGE:     method terminate
CONSOLE MESSAGE: interface WorkletAnimationEffect
CONSOLE MESSAGE:     attribute @@toStringTag
CONSOLE MESSAGE:     getter localTime
CONSOLE MESSAGE:     method constructor
CONSOLE MESSAGE:     method getComputedTiming
CONSOLE MESSAGE:     method getTiming
CONSOLE MESSAGE:     setter localTime
CONSOLE MESSAGE: interface WorkletGlobalScope
CONSOLE MESSAGE:     attribute @@toStringTag
CONSOLE MESSAGE:     method constructor
CONSOLE MESSAGE: interface WorkletGroupEffect
CONSOLE MESSAGE:     attribute @@toStringTag
CONSOLE MESSAGE:     method constructor
CONSOLE MESSAGE:     method getChildren
CONSOLE MESSAGE: interface WritableStream
CONSOLE MESSAGE:     attribute @@toStringTag
CONSOLE MESSAGE:     getter locked
CONSOLE MESSAGE:     method abort
CONSOLE MESSAGE:     method close
CONSOLE MESSAGE:     method constructor
CONSOLE MESSAGE:     method getWriter
CONSOLE MESSAGE: interface WritableStreamDefaultController
CONSOLE MESSAGE:     attribute @@toStringTag
CONSOLE MESSAGE:     getter signal
CONSOLE MESSAGE:     method constructor
CONSOLE MESSAGE:     method error
CONSOLE MESSAGE: interface WritableStreamDefaultWriter
CONSOLE MESSAGE:     attribute @@toStringTag
CONSOLE MESSAGE:     getter closed
CONSOLE MESSAGE:     getter desiredSize
CONSOLE MESSAGE:     getter ready
CONSOLE MESSAGE:     method abort
CONSOLE MESSAGE:     method close
CONSOLE MESSAGE:     method constructor
CONSOLE MESSAGE:     method releaseLock
CONSOLE MESSAGE:     method write
CONSOLE MESSAGE: [NAMESPACES]
CONSOLE MESSAGE: namespace console
CONSOLE MESSAGE:     attribute @@toStringTag
CONSOLE MESSAGE:     method assert
CONSOLE MESSAGE:     method clear
CONSOLE MESSAGE:     method context
CONSOLE MESSAGE:     method count
CONSOLE MESSAGE:     method countReset
CONSOLE MESSAGE:     method createTask
CONSOLE MESSAGE:     method debug
CONSOLE MESSAGE:     method dir
CONSOLE MESSAGE:     method dirxml
CONSOLE MESSAGE:     method error
CONSOLE MESSAGE:     method group
CONSOLE MESSAGE:     method groupCollapsed
CONSOLE MESSAGE:     method groupEnd
CONSOLE MESSAGE:     method info
CONSOLE MESSAGE:     method log
CONSOLE MESSAGE:     method profile
CONSOLE MESSAGE:     method profileEnd
CONSOLE MESSAGE:     method table
CONSOLE MESSAGE:     method time
CONSOLE MESSAGE:     method timeEnd
CONSOLE MESSAGE:     method timeLog
CONSOLE MESSAGE:     method timeStamp
CONSOLE MESSAGE:     method trace
CONSOLE MESSAGE:     method warn
CONSOLE MESSAGE: [GLOBAL OBJECT]
CONSOLE MESSAGE:     attribute globalThis
CONSOLE MESSAGE:     method gc
CONSOLE MESSAGE:     method registerAnimator
CONSOLE MESSAGE: [INTERFACES]
CONSOLE MESSAGE: interface AnimationWorkletGlobalScope : WorkletGlobalScope
CONSOLE MESSAGE:     attribute @@toStringTag
CONSOLE MESSAGE:     method constructor
CONSOLE MESSAGE: interface ByteLengthQueuingStrategy
CONSOLE MESSAGE:     attribute @@toStringTag
CONSOLE MESSAGE:     getter highWaterMark
CONSOLE MESSAGE:     getter size
CONSOLE MESSAGE:     method constructor
CONSOLE MESSAGE: interface CountQueuingStrategy
CONSOLE MESSAGE:     attribute @@toStringTag
CONSOLE MESSAGE:     getter highWaterMark
CONSOLE MESSAGE:     getter size
CONSOLE MESSAGE:     method constructor
CONSOLE MESSAGE: interface ReadableByteStreamController
CONSOLE MESSAGE:     attribute @@toStringTag
CONSOLE MESSAGE:     getter byobRequest
CONSOLE MESSAGE:     getter desiredSize
CONSOLE MESSAGE:     method close
CONSOLE MESSAGE:     method constructor
CONSOLE MESSAGE:     method enqueue
CONSOLE MESSAGE:     method error
CONSOLE MESSAGE: interface ReadableStream
CONSOLE MESSAGE:     attribute @@toStringTag
CONSOLE MESSAGE:     getter locked
CONSOLE MESSAGE:     method cancel
CONSOLE MESSAGE:     method constructor
CONSOLE MESSAGE:     method getReader
CONSOLE MESSAGE:     method pipeThrough
CONSOLE MESSAGE:     method pipeTo
CONSOLE MESSAGE:     method tee
CONSOLE MESSAGE: interface ReadableStreamBYOBReader
CONSOLE MESSAGE:     attribute @@toStringTag
CONSOLE MESSAGE:     getter closed
CONSOLE MESSAGE:     method cancel
CONSOLE MESSAGE:     method constructor
CONSOLE MESSAGE:     method read
CONSOLE MESSAGE:     method releaseLock
CONSOLE MESSAGE: interface ReadableStreamBYOBRequest
CONSOLE MESSAGE:     attribute @@toStringTag
CONSOLE MESSAGE:     getter view
CONSOLE MESSAGE:     method constructor
CONSOLE MESSAGE:     method respond
CONSOLE MESSAGE:     method respondWithNewView
CONSOLE MESSAGE: interface ReadableStreamDefaultController
CONSOLE MESSAGE:     attribute @@toStringTag
CONSOLE MESSAGE:     getter desiredSize
CONSOLE MESSAGE:     method close
CONSOLE MESSAGE:     method constructor
CONSOLE MESSAGE:     method enqueue
CONSOLE MESSAGE:     method error
CONSOLE MESSAGE: interface ReadableStreamDefaultReader
CONSOLE MESSAGE:     attribute @@toStringTag
CONSOLE MESSAGE:     getter closed
CONSOLE MESSAGE:     method cancel
CONSOLE MESSAGE:     method constructor
CONSOLE MESSAGE:     method read
CONSOLE MESSAGE:     method releaseLock
CONSOLE MESSAGE: interface TransformStream
CONSOLE MESSAGE:     attribute @@toStringTag
CONSOLE MESSAGE:     getter readable
CONSOLE MESSAGE:     getter writable
CONSOLE MESSAGE:     method constructor
CONSOLE MESSAGE: interface TransformStreamDefaultController
CONSOLE MESSAGE:     attribute @@toStringTag
CONSOLE MESSAGE:     getter desiredSize
CONSOLE MESSAGE:     method constructor
CONSOLE MESSAGE:     method enqueue
CONSOLE MESSAGE:     method error
CONSOLE MESSAGE:     method terminate
CONSOLE MESSAGE: interface WorkletAnimationEffect
CONSOLE MESSAGE:     attribute @@toStringTag
CONSOLE MESSAGE:     getter localTime
CONSOLE MESSAGE:     method constructor
CONSOLE MESSAGE:     method getComputedTiming
CONSOLE MESSAGE:     method getTiming
CONSOLE MESSAGE:     setter localTime
CONSOLE MESSAGE: interface WorkletGlobalScope
CONSOLE MESSAGE:     attribute @@toStringTag
CONSOLE MESSAGE:     method constructor
CONSOLE MESSAGE: interface WorkletGroupEffect
CONSOLE MESSAGE:     attribute @@toStringTag
CONSOLE MESSAGE:     method constructor
CONSOLE MESSAGE:     method getChildren
CONSOLE MESSAGE: interface WritableStream
CONSOLE MESSAGE:     attribute @@toStringTag
CONSOLE MESSAGE:     getter locked
CONSOLE MESSAGE:     method abort
CONSOLE MESSAGE:     method close
CONSOLE MESSAGE:     method constructor
CONSOLE MESSAGE:     method getWriter
CONSOLE MESSAGE: interface WritableStreamDefaultController
CONSOLE MESSAGE:     attribute @@toStringTag
CONSOLE MESSAGE:     getter signal
CONSOLE MESSAGE:     method constructor
CONSOLE MESSAGE:     method error
CONSOLE MESSAGE: interface WritableStreamDefaultWriter
CONSOLE MESSAGE:     attribute @@toStringTag
CONSOLE MESSAGE:     getter closed
CONSOLE MESSAGE:     getter desiredSize
CONSOLE MESSAGE:     getter ready
CONSOLE MESSAGE:     method abort
CONSOLE MESSAGE:     method close
CONSOLE MESSAGE:     method constructor
CONSOLE MESSAGE:     method releaseLock
CONSOLE MESSAGE:     method write
CONSOLE MESSAGE: [NAMESPACES]
CONSOLE MESSAGE: namespace console
CONSOLE MESSAGE:     attribute @@toStringTag
CONSOLE MESSAGE:     method assert
CONSOLE MESSAGE:     method clear
CONSOLE MESSAGE:     method context
CONSOLE MESSAGE:     method count
CONSOLE MESSAGE:     method countReset
CONSOLE MESSAGE:     method createTask
CONSOLE MESSAGE:     method debug
CONSOLE MESSAGE:     method dir
CONSOLE MESSAGE:     method dirxml
CONSOLE MESSAGE:     method error
CONSOLE MESSAGE:     method group
CONSOLE MESSAGE:     method groupCollapsed
CONSOLE MESSAGE:     method groupEnd
CONSOLE MESSAGE:     method info
CONSOLE MESSAGE:     method log
CONSOLE MESSAGE:     method profile
CONSOLE MESSAGE:     method profileEnd
CONSOLE MESSAGE:     method table
CONSOLE MESSAGE:     method time
CONSOLE MESSAGE:     method timeEnd
CONSOLE MESSAGE:     method timeLog
CONSOLE MESSAGE:     method timeStamp
CONSOLE MESSAGE:     method trace
CONSOLE MESSAGE:     method warn
CONSOLE MESSAGE: [GLOBAL OBJECT]
CONSOLE MESSAGE:     attribute globalThis
CONSOLE MESSAGE:     method gc
CONSOLE MESSAGE:     method registerAnimator
